# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Product": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"url_photo": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"old_price": {
						"type": "number",
						"format": "double"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"items": {
							"$ref": "#/components/schemas/Category"
						},
						"type": "array"
					},
					"home": {
						"items": {
							"$ref": "#/components/schemas/Home"
						},
						"type": "array"
					},
					"markets": {
						"items": {
							"$ref": "#/components/schemas/Market"
						},
						"type": "array"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/Order"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"url_photo",
					"price",
					"old_price",
					"discount",
					"category",
					"home",
					"markets",
					"orders",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"url_photo": {
						"type": "string"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"url_photo",
					"products",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Home": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"type_product": {
						"type": "string"
					},
					"product": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"type_product",
					"product",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Order": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"total_value": {
						"type": "number",
						"format": "double"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"market": {
						"$ref": "#/components/schemas/Market"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"address": {
						"$ref": "#/components/schemas/Address"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"total_value",
					"products",
					"market",
					"user",
					"address",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"City": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"state": {
						"$ref": "#/components/schemas/State"
					},
					"adresses": {
						"items": {
							"$ref": "#/components/schemas/Address"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"state",
					"adresses",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"State": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"uf": {
						"type": "string"
					},
					"cities": {
						"items": {
							"$ref": "#/components/schemas/City"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"uf",
					"cities",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Address": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"street": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"cep": {
						"type": "string"
					},
					"complement": {
						"type": "string"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/Order"
						},
						"type": "array"
					},
					"market": {
						"$ref": "#/components/schemas/Market"
					},
					"user": {
						"$ref": "#/components/schemas/Market"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"street",
					"number",
					"district",
					"cep",
					"complement",
					"city",
					"orders",
					"market",
					"user",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Market": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"url_photo": {
						"type": "string"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/Order"
						},
						"type": "array"
					},
					"address": {
						"$ref": "#/components/schemas/Address"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"url_photo",
					"products",
					"orders",
					"address",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password_hash": {
						"type": "string"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/Order"
						},
						"type": "array"
					},
					"address": {
						"$ref": "#/components/schemas/Address"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password_hash",
					"orders",
					"address",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IOrderPayload": {
				"properties": {
					"productsIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"marketId": {
						"type": "number",
						"format": "double"
					},
					"addressId": {
						"type": "number",
						"format": "double"
					},
					"user": {}
				},
				"required": [
					"productsIds",
					"marketId",
					"addressId",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.1.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/order/{id}": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Order"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/order": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrderPayload"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}